examples/flask-form/requirements.txt
flask
flask-wtf
flask-sqlalchemy
sqlalchemy-media
